refactor: dashboard, house, issue, image
