--- frontend/src/components/House/index.js
@@ -1,5 +1,0 @@
-import House from './House';
-import HouseList from './HouseList';
-import HouseForm from './HouseForm';
-
-export {House, HouseList, HouseForm};
