--- frontend/src/components/Issue/index.js
@@ -1,4 +1,0 @@
-import FilteredIssueList from './FilteredIssueList';
-import IssueForm from './IssueForm';
-
-export {FilteredIssueList, IssueForm};
