+++ frontend/src/containers/House/HouseCard.js
