+++ frontend/src/containers/House/index.js
@@ -1,3 +1,5 @@
 import House from './House';
+import HouseList from './HouseList';
 
 export default House;
+export {HouseList};
