+++ frontend/src/containers/House/styles.js
