+++ frontend/src/containers/Image/ImageRemovable.js
