+++ frontend/src/containers/Image/styles.js
