+++ frontend/src/containers/Issue/index.js
@@ -1,3 +1,5 @@
 import Issue from './Issue';
+import FilteredIssueList from './FilteredIssueList';
 
 export default Issue;
+export {FilteredIssueList};
