+++ frontend/src/containers/Uploader/CloseButton.js
