+++ frontend/src/containers/Uploader/ImageDropper.js
